+++ concordium-std-derive/src/lib.rs
@@ -46,12 +46,15 @@ struct OptionalArguments {
     pub(crate) enable_logger:     bool,
     /// The function is a low-level one, with direct access to contract memory.
     pub(crate) low_level:         bool,
-    /// Which type, if any, is the parameter type of the contract.
+    /// Which type, if any, is the parameter type of the function.
     /// This is used when generating schemas.
     pub(crate) parameter:         Option<syn::LitStr>,
-    /// Which type, if any, is the return value of the contract.
+    /// Which type, if any, is the return value of the function.
     /// This is used when generating schemas.
     pub(crate) return_value:      Option<syn::LitStr>,
+    /// Which type, if any, is the error type of the function.
+    /// This is used when generating schemas.
+    pub(crate) error:             Option<syn::LitStr>,
     /// If enabled, the function has access to cryptographic primitives.
     pub(crate) crypto_primitives: bool,
 }
@@ -234,6 +237,7 @@ const INIT_ATTRIBUTE_PAYABLE: &str = "payable";
 const INIT_ATTRIBUTE_ENABLE_LOGGER: &str = "enable_logger";
 const INIT_ATTRIBUTE_LOW_LEVEL: &str = "low_level";
 const INIT_ATTRIBUTE_RETURN_VALUE: &str = "return_value";
+const INIT_ATTRIBUTE_ERROR: &str = "error";
 const INIT_ATTRIBUTE_CRYPTO_PRIMITIVES: &str = "crypto_primitives";
 
 fn parse_init_attributes<'a, I: IntoIterator<Item = &'a Meta>>(
@@ -259,6 +263,7 @@ fn parse_init_attributes<'a, I: IntoIterator<Item = &'a Meta>>(
             "The 'return_value' attribute is currently not supported for init methods.",
         ));
     }
+    let error = attributes.extract_value(INIT_ATTRIBUTE_ERROR);
     let crypto_primitives = attributes.extract_flag(INIT_ATTRIBUTE_CRYPTO_PRIMITIVES).is_some();
 
     // Make sure that there are no unrecognized attributes. These would typically be
@@ -273,6 +278,7 @@ fn parse_init_attributes<'a, I: IntoIterator<Item = &'a Meta>>(
             low_level,
             parameter,
             return_value: None, // Return values are currently not supported on init methods.
+            error,
             crypto_primitives,
         },
     })
@@ -282,6 +288,7 @@ fn parse_init_attributes<'a, I: IntoIterator<Item = &'a Meta>>(
 
 const RECEIVE_ATTRIBUTE_PARAMETER: &str = "parameter";
 const RECEIVE_ATTRIBUTE_RETURN_VALUE: &str = "return_value";
+const RECEIVE_ATTRIBUTE_ERROR: &str = "error";
 const RECEIVE_ATTRIBUTE_CONTRACT: &str = "contract";
 const RECEIVE_ATTRIBUTE_NAME: &str = "name";
 const RECEIVE_ATTRIBUTE_FALLBACK: &str = "fallback";
@@ -302,6 +309,7 @@ fn parse_receive_attributes<'a, I: IntoIterator<Item = &'a Meta>>(
     let parameter: Option<syn::LitStr> = attributes.extract_value(RECEIVE_ATTRIBUTE_PARAMETER);
     let return_value: Option<syn::LitStr> =
         attributes.extract_value(RECEIVE_ATTRIBUTE_RETURN_VALUE);
+    let error: Option<syn::LitStr> = attributes.extract_value(RECEIVE_ATTRIBUTE_ERROR);
     let payable = attributes.extract_flag(RECEIVE_ATTRIBUTE_PAYABLE).is_some();
     let enable_logger = attributes.extract_flag(RECEIVE_ATTRIBUTE_ENABLE_LOGGER).is_some();
     let low_level = attributes.extract_flag(RECEIVE_ATTRIBUTE_LOW_LEVEL);
@@ -349,6 +357,7 @@ fn parse_receive_attributes<'a, I: IntoIterator<Item = &'a Meta>>(
                 low_level: low_level.is_some(),
                 parameter,
                 return_value,
+                error,
                 crypto_primitives,
             },
             mutable: mutable.is_some(), /* This is also optional, but does not belong in
@@ -366,6 +375,7 @@ fn parse_receive_attributes<'a, I: IntoIterator<Item = &'a Meta>>(
                         low_level: low_level.is_some(),
                         parameter,
                         return_value,
+                        error,
                         crypto_primitives,
                     },
                     mutable: mutable.is_some(), /* TODO: This is also optional, but does not
@@ -468,7 +478,7 @@ fn contains_attribute<'a, I: IntoIterator<Item = &'a Meta>>(iter: I, name: &str)
 /// ```
 ///
 /// ## `parameter="<Param>"`: Generate schema for parameter
-/// To make schema generation to include the parameter for this function, add
+/// To make schema generation include the parameter for this function, add
 /// the attribute `parameter` and set it equal to a string literal containing
 /// the name of the type used for the parameter. The parameter type must
 /// implement the SchemaType trait, which for most cases can be derived
@@ -480,6 +490,22 @@ fn contains_attribute<'a, I: IntoIterator<Item = &'a Meta>>(iter: I, name: &str)
 /// struct MyParam { ... }
 ///
 /// #[init(contract = "my_contract", parameter = "MyParam")]
+/// ```
+///
+/// ## `error="<Error>"`: Generate schema for error
+/// To make schema generation include the error for this function, add
+/// the attribute `error` and set it equal to a string literal containing
+/// the name of the type used for the error. The error type must
+/// implement the SchemaType trait, which for most cases can be derived
+/// automatically.
+///
+/// ### Example
+/// ```ignore
+/// #[derive(SchemaType)]
+/// enum MyError { ... }
+///
+/// #[init(contract = "my_contract", parameter = "MyError")]
+/// fn some_init(ctx: &impl HasInitContext, state: &mut impl HasStateApi) -> Result<(), MyError> {...}
 /// ```
 ///
 /// ## `crypto_primitives`: Function can access cryptographic primitives
@@ -600,9 +626,11 @@ fn init_worker(attr: TokenStream, item: TokenStream) -> syn::Result<TokenStream>
     // attribute is set.
     let parameter_option = init_attributes.optional.parameter;
     let return_value_option = None; // Return values are currently not supported on init.
+    let error_option = init_attributes.optional.error;
     out.extend(contract_function_schema_tokens(
         parameter_option,
         return_value_option,
+        error_option,
         rust_export_fn_name,
         wasm_export_fn_name,
     )?);
@@ -750,6 +778,25 @@ fn init_worker(attr: TokenStream, item: TokenStream) -> syn::Result<TokenStream>
 ///    ctx: &impl HasReceiveContext,
 ///    host: &HasHost<MyState, StateApiType = S>,
 /// ) -> ReceiveResult<MyReturnValue> {...}
+/// ```
+///
+/// ## `error="<Error>"`: Generate schema for error
+/// To make schema generation include the error for this function, add
+/// the attribute `error` and set it equal to a string literal containing
+/// the type used for the error. The error type must
+/// implement the SchemaType trait, which for most cases can be derived
+/// automatically.
+///
+/// ### Example
+/// ```ignore
+/// #[derive(SchemaType)]
+/// enum MyError { ... }
+///
+/// #[receive(contract = "my_contract", name = "some_receive", error = "MyError")]
+/// fn contract_receive<S: HasStateApi>(
+///     ctx: &impl HasReceiveContext,
+///     host: &HasHost<MyState, StateApiType = S>,
+/// ) -> Result<A, MyError> {...}
 /// ```
 ///
 /// ## `fallback`: Create a fallback entrypoint.
@@ -946,9 +993,11 @@ fn receive_worker(attr: TokenStream, item: TokenStream) -> syn::Result<TokenStre
     // attribute is set.
     let parameter_option = receive_attributes.optional.parameter;
     let return_value_option = receive_attributes.optional.return_value;
+    let error_option = receive_attributes.optional.error;
     out.extend(contract_function_schema_tokens(
         parameter_option,
         return_value_option,
+        error_option,
         rust_export_fn_name,
         wasm_export_fn_name,
     )?);
@@ -1003,32 +1052,67 @@ fn contract_function_optional_args_tokens(
 fn contract_function_schema_tokens(
     parameter_option: Option<syn::LitStr>,
     return_value_option: Option<syn::LitStr>,
+    error_option: Option<syn::LitStr>,
     rust_name: syn::Ident,
     wasm_name: String,
 ) -> syn::Result<proc_macro2::TokenStream> {
-    let construct_schema_bytes = match (parameter_option, return_value_option) {
-        (Some(parameter_ty), Some(return_value_ty)) => {
+    let construct_schema_bytes = match (parameter_option, return_value_option, error_option) {
+        (Some(parameter_ty), None, None) => {
+            let parameter_ty = parameter_ty.parse::<syn::Type>()?;
+            Some(quote! {
+            let parameter = <#parameter_ty as schema::SchemaType>::get_type();
+            let schema_bytes = concordium_std::to_bytes(&schema::FunctionV2::Param(parameter));})
+        }
+        (None, Some(return_value_ty), None) => {
+            let return_value_ty = return_value_ty.parse::<syn::Type>()?;
+            Some(quote! {
+            let return_value = <#return_value_ty as schema::SchemaType>::get_type();
+            let schema_bytes =
+                concordium_std::to_bytes(&schema::FunctionV2::Rv(return_value));     })
+        }
+        (Some(parameter_ty), Some(return_value_ty), None) => {
             let parameter_ty = parameter_ty.parse::<syn::Type>()?;
             let return_value_ty = return_value_ty.parse::<syn::Type>()?;
             Some(quote! {
-                let parameter = <#parameter_ty as schema::SchemaType>::get_type();
-                let return_value = <#return_value_ty as schema::SchemaType>::get_type();
-                let schema_bytes = concordium_std::to_bytes(&schema::Function::Both { parameter, return_value });
-            })
+                    let parameter = <#parameter_ty as schema::SchemaType>::get_type();
+                    let return_value = <#return_value_ty as schema::SchemaType>::get_type();
+                    let schema_bytes = concordium_std::to_bytes(&schema::FunctionV2::ParamRv {
+            parameter, return_value });     })
         }
-        (Some(parameter_ty), None) => {
+        (None, None, Some(error_ty)) => {
+            let error_ty = error_ty.parse::<syn::Type>()?;
+            Some(quote! {
+            let error = <#error_ty as schema::SchemaType>::get_type();
+            let schema_bytes = concordium_std::to_bytes(&schema::FunctionV2::Error(error));     })
+        }
+        (Some(parameter_ty), None, Some(error_ty)) => {
             let parameter_ty = parameter_ty.parse::<syn::Type>()?;
+            let error_ty = error_ty.parse::<syn::Type>()?;
             Some(quote! {
-                let parameter = <#parameter_ty as schema::SchemaType>::get_type();
-                let schema_bytes = concordium_std::to_bytes(&schema::Function::Parameter(parameter));
-            })
+                    let parameter = <#parameter_ty as schema::SchemaType>::get_type();
+                    let error = <#error_ty as schema::SchemaType>::get_type();
+                    let schema_bytes = concordium_std::to_bytes(&schema::FunctionV2::ParamError {
+            parameter, error });     })
         }
-        (None, Some(return_value_ty)) => {
+        (None, Some(return_value_ty), Some(error_ty)) => {
             let return_value_ty = return_value_ty.parse::<syn::Type>()?;
+            let error_ty = error_ty.parse::<syn::Type>()?;
             Some(quote! {
-                let return_value = <#return_value_ty as schema::SchemaType>::get_type();
-                let schema_bytes = concordium_std::to_bytes(&schema::Function::ReturnValue(return_value));
-            })
+                    let return_value = <#return_value_ty as schema::SchemaType>::get_type();
+                    let error = <#error_ty as schema::SchemaType>::get_type();
+                    let schema_bytes = concordium_std::to_bytes(&schema::FunctionV2::RvError {
+            return_value, error });     })
+        }
+        (Some(parameter_ty), Some(return_value_ty), Some(error_ty)) => {
+            let parameter_ty = parameter_ty.parse::<syn::Type>()?;
+            let return_value_ty = return_value_ty.parse::<syn::Type>()?;
+            let error_ty = error_ty.parse::<syn::Type>()?;
+            Some(quote! {
+                    let parameter = <#parameter_ty as schema::SchemaType>::get_type();
+                    let return_value = <#return_value_ty as schema::SchemaType>::get_type();
+                    let error = <#error_ty as schema::SchemaType>::get_type();
+                    let schema_bytes = concordium_std::to_bytes(&schema::FunctionV2::ParamRvError {
+            parameter, return_value, error });     })
         }
         _ => None,
     };
@@ -1054,6 +1138,7 @@ fn contract_function_schema_tokens(
 fn contract_function_schema_tokens(
     _parameter_option: Option<syn::LitStr>,
     _return_value_option: Option<syn::LitStr>,
+    _error_option: Option<syn::LitStr>,
     _rust_name: syn::Ident,
     _wasm_name: String,
 ) -> syn::Result<proc_macro2::TokenStream> {
@@ -1559,16 +1644,10 @@ const RESERVED_ERROR_CODES: i32 = i32::MIN + 100;
 /// Creating custom enums for error types can provide meaningful error messages
 /// to the user of the smart contract.
 ///
-/// For enums where at least one variant has a field, the enum *must* implement
+/// When a contract function rejects, the enum is serialized and returned along
+/// with the error code. The serialization means that the enum *must* implement
 /// [`Serial`](../concordium_contracts_common/trait.Serial.html) if [`Reject`]
 /// is to be derived.
-///
-/// For variants with fields, the enum is serialized into the buffer used in
-/// the return value, which is always [`Some`][std::option::Option::Some].
-/// The contents of the buffer are returned by the contract along with the error
-/// code.
-/// For fieldless variants the return value is always
-/// [`None`][std::option::Option::None] and only the error code is returned.
 ///
 /// The conversion will map the first variant to error code -1, second to -2,
 /// etc.
@@ -1579,7 +1658,7 @@ const RESERVED_ERROR_CODES: i32 = i32::MIN + 100;
 /// enum MyError {
 ///     IllegalState, // receives error code -1
 ///     WrongSender, // receives error code -2
-///     TimeExpired(time: Timestamp), // receives error code -3 (enums with fields must implement Serial)
+///     TimeExpired(time: Timestamp), // receives error code -3
 ///     ...
 /// }
 /// ```
@@ -1616,8 +1695,8 @@ fn reject_derive_worker(input: TokenStream) -> syn::Result<TokenStream> {
     };
 
     let variant_error_conversions = generate_variant_error_conversions(enum_data, enum_ident)?;
-    let enum_var_ident = format_ident!("{}", "e");
-    let variant_matches = generate_variant_matches(enum_data, enum_ident, &enum_var_ident);
+    let buf_var_ident = format_ident!("{}", "buf");
+    let variant_matches = generate_variant_matches(enum_data, enum_ident, &buf_var_ident);
 
     let gen = quote! {
         /// The from implementation maps the first variant to -1, second to -2, etc.
@@ -1626,8 +1705,10 @@ fn reject_derive_worker(input: TokenStream) -> syn::Result<TokenStream> {
         #[automatically_derived]
         impl From<#enum_ident> for Reject {
             #[inline(always)]
-            fn from(#enum_var_ident: #enum_ident) -> Self {
-                match &#enum_var_ident {
+            fn from(e: #enum_ident) -> Self {
+                let mut #buf_var_ident = Vec::new();
+                concordium_std::Serial::serial(&e, &mut #buf_var_ident).unwrap_abort();
+                match &e {
                    #variant_matches
                 }
             }
@@ -1638,15 +1719,14 @@ fn reject_derive_worker(input: TokenStream) -> syn::Result<TokenStream> {
     Ok(gen.into())
 }
 
-/// Generate the cases for matching on the enum. For variants with fields, the
-/// whole enum type is serialized into a buffer that is included in the return
-/// value. This means that enums with fields need to implement `Serial` in order
-/// for deriving `Reject` to work. For fieldless variants, the return value is
-/// always `None`.
+/// Generate the cases for matching on the enum.
+/// The error codes for variants start at -1 and go downwards.
+/// The whole enum is serialized and included in the return_value field, which,
+/// thus, is always `Some`.
 fn generate_variant_matches(
     enum_data: &DataEnum,
     enum_name: &syn::Ident,
-    enum_var_ident: &syn::Ident,
+    buf_var_ident: &syn::Ident,
 ) -> proc_macro2::TokenStream {
     let mut match_cases = proc_macro2::TokenStream::new();
     for (index, variant) in enum_data.variants.iter().enumerate() {
@@ -1656,11 +1736,7 @@ fn generate_variant_matches(
                 match_cases.extend(quote! {
                     #enum_name::#variant_ident{..} => Reject {
                         error_code: unsafe { num::NonZeroI32::new_unchecked(-(#index as i32) - 1) },
-                        return_value: {
-                            let mut buf = Vec::new();
-                            concordium_std::Serial::serial(&#enum_var_ident, &mut buf).unwrap_abort();
-                            Some(buf)
-                        },
+                        return_value: Some(#buf_var_ident),
                     },
                 });
             }
@@ -1668,11 +1744,7 @@ fn generate_variant_matches(
                 match_cases.extend(quote! {
                     #enum_name::#variant_ident(..) => Reject {
                         error_code: unsafe { num::NonZeroI32::new_unchecked(-(#index as i32) - 1) },
-                        return_value: {
-                            let mut buf = Vec::new();
-                            concordium_std::Serial::serial(&#enum_var_ident, &mut buf).unwrap_abort();
-                            Some(buf)
-                        },
+                        return_value: Some(#buf_var_ident),
                     },
                 });
             }
@@ -1680,7 +1752,7 @@ fn generate_variant_matches(
                 match_cases.extend(quote! {
                     #enum_name::#variant_ident => Reject {
                         error_code: unsafe { num::NonZeroI32::new_unchecked(-(#index as i32) - 1) },
-                        return_value: None,
+                        return_value: Some(#buf_var_ident),
                     },
                 });
             }